MVVM Refactorization
