+++ app/src/main/java/com/example/pokemonslist/data/remote/PokeApiService.kt
@@ -1,7 +1,7 @@
-package com.example.pokemonslist.network
+package com.example.pokemonslist.data.remote
 
-import com.example.pokemonslist.data.PokemonListResponse
-import com.example.pokemonslist.data.PokemonResponse
+import com.example.pokemonslist.data.repository.PokemonListResponse
+import com.example.pokemonslist.data.repository.PokemonResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 import retrofit2.http.Path
