+++ app/src/main/java/com/example/pokemonslist/data/repository/PokemonListResponse.kt
@@ -1,4 +1,4 @@
-package com.example.pokemonslist.data
+package com.example.pokemonslist.data.repository
 
 data class PokemonListResponse (
     val results: List<PokemonListItem>
