+++ app/src/main/java/com/example/pokemonslist/ui/viewmodel/PokemonViewModel.kt
@@ -3,9 +3,8 @@ package com.example.pokemonslist.ui.viewmodel
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.pokemonslist.data.PokemonResponse
-import com.example.pokemonslist.network.PokeApiService
-import com.example.pokemonslist.network.RetrofitClient
+import com.example.pokemonslist.data.repository.MainRepository
+import com.example.pokemonslist.data.repository.PokemonResponse
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -19,12 +18,10 @@ sealed interface PokemonListUiState {
     data object Loading : PokemonListUiState
 }
 
-class PokemonViewModel : ViewModel() {
+class PokemonViewModel(private val repository: MainRepository) : ViewModel() {
 
     private val _uiState = MutableStateFlow<PokemonListUiState>(PokemonListUiState.Loading)
     val uiState: StateFlow<PokemonListUiState> = _uiState.asStateFlow()
-
-    private val pokeApiService: PokeApiService = RetrofitClient.instance
 
     init {
         fetchPokemons()
@@ -34,7 +31,7 @@ class PokemonViewModel : ViewModel() {
         _uiState.value = PokemonListUiState.Loading
         viewModelScope.launch {
             try {
-                val pokemonListItems = pokeApiService.getPokemonList(limit = 100, offset = 0).results
+                val pokemonListItems = repository.getPokemonList(limit = 100, offset = 0).results
 
                 val detailedPokemons = pokemonListItems.map { listItem ->
                     async {
@@ -42,7 +39,7 @@ class PokemonViewModel : ViewModel() {
                         val urlParts = listItem.url.trimEnd('/').split('/')
                         val pokemonId = urlParts.last()
                         Log.d("PokemonViewModel", "Extracted ID: $pokemonId for URL: ${listItem.url}")
-                        pokeApiService.getPokemon(pokemonId)
+                        repository.getPokemon(pokemonId)
                     }
                 }.awaitAll()
